T-007: Decompose MainViewModel into feature-specific ViewModels
Split the monolithic MainViewModel (~1420 lines) into 5 focused ViewModels:
- BleViewModel: BLE connection lifecycle, data pipeline, device management
- SessionViewModel: Device write commands (heater, temp, boost, settings)
- HistoryViewModel: Session list, filtering, sorting, analytics, export
- BatteryViewModel: Battery insights, charge cycles, card UI state
- SettingsViewModel: User preferences (day start, retention, capsule weight)

Updated all fragments and MainActivity to use the new ViewModels via
activityViewModels(). Cross-VM state sync (activeDevice, dayStartHour)
is handled in MainActivity. MainViewModel retained as empty Hilt shell
for backward compatibility.
